+++ Syclight/src/Syclight/LayerStack.h
@@ -12,7 +12,7 @@ namespace syc
 	private:
 		std::vector<Layer*> m_Layers;
 		/*std::vector<Layer*>::iterator m_LayerInsert;*/
-		uint16 m_LayerInsertIndex = 0;
+		uint32 m_LayerInsertIndex = 0;
 
 	public:
 		LayerStack();
